Move builder to its own class
